Updated Frontend and Backend (using GET Request To Update Highscore because of CSRF trouble)

+++ shootingRangeBackend/user_auth/models.py
@@ -5,8 +5,8 @@ from django.contrib.auth.models import User
 
 class UserProfile(models.Model):
     user = models.OneToOneField(User, on_delete=models.CASCADE)
-    easy_high_score = models.IntegerField(default=0)
-    medium_high_score = models.IntegerField(default=0)
-    hard_high_score = models.IntegerField(default=0)
+    easy_high_score = models.FloatField(default=0)
+    medium_high_score = models.FloatField(default=0)
+    hard_high_score = models.FloatField(default=0)
     def __str__(self):
         return f'{self.user.username} Profile'
